Pass event handler to the element that we want to monitor, which is the <input /> element
All input elements emit a 'change' event. 'change' is the name of the event. Whenever a user interacts with an input
element by typing something into it, this triggers the 'change' event. In order to use regular html events like'change'
all you have to do is within the element you're monitoring type the word 'on' and then the name of the event 'change'
and then set it equals to a reference to our event handler method (onInputChange) within curly braces. So you end up with this:
onChange={this.onInputChange}