Add the 'event_list_item' component
So the event_list_item component is now going to represent an event as it appears in a list. So now
we will change the event_detail component to be used for whenever a user clicks to view more info on
one individual event. The event_detail component will be used for the event landing pages.