Restructure the event_detail component to make usable for landing pages
The event_detail component is going to be the html that shows up on the event landing pages. One thing to notice
is that I'm using an if statement to first make sure that the event object/prop is not empty, but if it is, to
render a 'loading' text until the object has data, which should trigger a re-rendering of the page by React.