Add the ability to select an event from the list of events
We added a function to the App component called 'onEventSelect'. The purpose of this function is to
take an event and then update the selectedEvent state. We pass the 'onEventSelect' function as a prop
down to the EventList component. EventList takes this prop and passes it down to the EventListItem component.
EventListItem takes the 'onEventSelect' prop and basically says 'whenever I get clicked, call that function
with the event that I was passed. This then updates the selectedEvent, thus updating which event we are
displaying as the main event on the homepage.