Give us the ability to search for an event by the exact title
Had to switch up how I was trying to accomplish this, but basically the App component now has the 'searchTerm' state, as well as a function
'onSearchChange' that is used to update the state whenever someone types into the search field. I also added an 'isSearched' function to the
event_list component that is used to filter out the event that's being searched from the list of events that are available.